fix: migrate-based DB init, order auth, decimal money, antiforgery cart
- Switch EnsureCreated() to Migrate() so the schema is migration-driven
- Require authentication on order history; filter orders in the query
- Add deterministic product pagination ordering
- Store monetary values as decimal (add MoneyToDecimal migration)
- Require POST and antiforgery tokens for cart mutations

+++ Data/Cart/ShoppingCart.cs
@@ -96,16 +96,16 @@ namespace webshop_owp.Data.Cart
             return ShoppingCartItems ?? (ShoppingCartItems = _context.ShoppingCartItems.Where(n => n.ShoppingCartId == ShoppingCartId).Include(n => n.Product).ToList());
         }
 
-        public double GetShoppingCartTotal()
+        public decimal GetShoppingCartTotal()
         {
             var items = _context.ShoppingCartItems.Where(n => n.ShoppingCartId == ShoppingCartId).Include(n => n.Product).ToList();
-            double total = 0;
+            decimal total = 0;
             foreach (var item in items)
             {
                 var price = item.Product.Price;
                 if (item.Product.DiscountPercentage > 0)
                 {
-                    price = price * (1 - item.Product.DiscountPercentage / 100.0);
+                    price = price * (1 - item.Product.DiscountPercentage / 100m);
                 }
                 total += price * item.Amount;
             }
