+++ Data/ViewModels/ShoppingCartVM.cs
@@ -5,6 +5,6 @@ namespace webshop_owp.Data.ViewModels
     public class ShoppingCartVM
     {
         public ShoppingCart ShoppingCart { get; set; }
-        public double ShoppingCartTotal { get; set; }
+        public decimal ShoppingCartTotal { get; set; }
     }
 }
